refactor: made folder structure profile lowercase
